requirements.txt
transformers>=4.37.0
torch>=2.0.0
accelerate>=0.21.0
sentencepiece
peft>=0.7.0
datasets>=2.14.0
evaluate>=0.4.0
tqdm
wandb
bitsandbytes>=0.41.0
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
APScheduler==3.10.1
requests==2.31.0 